@@ -50,136 +50,135 @@
     "OperaMini all"
   ],
   "devDependencies": {
-    "@babel/cli": "^7.16.0",
-    "@babel/core": "^7.16.0",
-    "@babel/plugin-proposal-class-properties": "^7.16.0",
-    "@babel/plugin-proposal-optional-chaining": "^7.16.0",
-    "@babel/preset-env": "^7.16.0",
-    "@babel/preset-react": "^7.16.0",
-    "@babel/preset-typescript": "^7.16.0",
-    "@cypress/webpack-preprocessor": "^5.11.1",
-    "@emotion/babel-plugin": "^11.9.2",
-    "@emotion/jest": "^11.9.1",
-    "@guardian/eslint-config-typescript": "^1.0.11",
-    "@guardian/eslint-plugin-source-foundations": "^5.0.1",
-    "@guardian/eslint-plugin-source-react-components": "^5.0.0",
-    "@guardian/node-riffraff-artifact": "^0.2.2",
-    "@guardian/prettier": "^1.0.0",
-    "@guardian/tsconfig": "^0.1.6",
-    "@storybook/addon-essentials": "^6.4.19",
-    "@storybook/builder-webpack5": "^6.4.19",
-    "@storybook/manager-webpack5": "^6.4.19",
-    "@storybook/react": "^6.4.19",
-    "@testing-library/cypress": "^8.0.2",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@types/assets-webpack-plugin": "^7.1.0",
-    "@types/aws4": "^1.5.1",
-    "@types/base-64": "^0.1.2",
-    "@types/bunyan": "^1.8.4",
-    "@types/color": "^3.0.0",
-    "@types/cookie-parser": "^1.4.1",
-    "@types/csurf": "^1.9.35",
-    "@types/express": "^4.11.1",
-    "@types/google.analytics": "^0.0.39",
-    "@types/helmet": "^0.0.37",
-    "@types/jest": "^26.0.9",
-    "@types/js-cookie": "^2.2.2",
-    "@types/lodash": "^4.14.168",
-    "@types/node": "^10.1.2",
-    "@types/node-fetch": "^1.6.9",
-    "@types/number-to-words": "^1.2.0",
-    "@types/prettier": "^2.4.2",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
-    "@types/react-test-renderer": "^17.0.2",
-    "@types/ua-parser-js": "^0.7.36",
-    "@types/url-parse": "^1.4.3",
-    "@types/uuid": "^8.3.4",
-    "@types/webpack": "^5.28.0",
-    "@types/webpack-node-externals": "^2.5.3",
-    "@types/yup": "^0.26.23",
-    "assets-webpack-plugin": "^7.1.1",
-    "babel-jest": "^27.3.1",
-    "babel-loader": "^8.0.6",
-    "babel-loader-exclude-node-modules-except": "^1.2.1",
-    "babel-plugin-lodash": "^3.3.4",
-    "babel-plugin-source-map-support": "^2.1.1",
-    "chromatic": "^6.4.1",
-    "copy-node-modules": "^1.0.4",
-    "copy-webpack-plugin": "^9.1.0",
-    "core-js": "^3.19.1",
+    "@babel/cli": "7.16.0",
+    "@babel/core": "7.17.5",
+    "@babel/plugin-proposal-class-properties": "7.16.0",
+    "@babel/plugin-proposal-optional-chaining": "7.16.0",
+    "@babel/preset-env": "7.16.0",
+    "@babel/preset-react": "7.16.7",
+    "@babel/preset-typescript": "7.16.0",
+    "@cypress/webpack-preprocessor": "5.11.1",
+    "@emotion/babel-plugin": "11.9.2",
+    "@emotion/jest": "11.9.1",
+    "@guardian/eslint-config-typescript": "1.0.11",
+    "@guardian/eslint-plugin-source-foundations": "11.0.0",
+    "@guardian/eslint-plugin-source-react-components": "15.0.0",
+    "@guardian/node-riffraff-artifact": "0.2.2",
+    "@guardian/prettier": "1.0.0",
+    "@guardian/tsconfig": "0.1.6",
+    "@storybook/addon-essentials": "6.5.8",
+    "@storybook/builder-webpack5": "6.5.8",
+    "@storybook/manager-webpack5": "6.5.8",
+    "@storybook/react": "6.5.8",
+    "@testing-library/cypress": "8.0.2",
+    "@testing-library/jest-dom": "5.16.1",
+    "@testing-library/react": "12.1.2",
+    "@types/assets-webpack-plugin": "7.1.0",
+    "@types/aws4": "1.5.1",
+    "@types/base-64": "0.1.3",
+    "@types/bunyan": "1.8.6",
+    "@types/color": "3.0.0",
+    "@types/cookie-parser": "1.4.1",
+    "@types/csurf": "1.9.35",
+    "@types/express": "4.16.1",
+    "@types/google.analytics": "0.0.39",
+    "@types/helmet": "0.0.37",
+    "@types/jest": "26.0.9",
+    "@types/js-cookie": "2.2.3",
+    "@types/lodash": "4.14.168",
+    "@types/node": "10.14.4",
+    "@types/node-fetch": "1.6.9",
+    "@types/number-to-words": "1.2.0",
+    "@types/prettier": "2.4.2",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.1",
+    "@types/react-test-renderer": "18.0.0",
+    "@types/url-parse": "1.4.3",
+    "@types/uuid": "8.3.4",
+    "@types/webpack": "5.28.0",
+    "@types/webpack-node-externals": "2.5.3",
+    "@types/yup": "0.26.23",
+    "assets-webpack-plugin": "7.1.1",
+    "babel-jest": "27.3.1",
+    "babel-loader": "8.0.6",
+    "babel-loader-exclude-node-modules-except": "1.2.1",
+    "babel-plugin-lodash": "3.3.4",
+    "babel-plugin-source-map-support": "2.1.1",
+    "chromatic": "6.4.1",
+    "copy-node-modules": "1.1.1",
+    "copy-webpack-plugin": "9.1.0",
+    "core-js": "3.19.1",
     "cypress": "8.5.0",
-    "cypress-plugin-stripe-elements": "^1.0.2",
-    "eslint": "^8.0.0",
-    "eslint-plugin-jest": "^27.0.4",
-    "eslint-plugin-react": "^7.30.0",
-    "fetch-mock": "^9.11.0",
-    "git-revision-webpack-plugin": "^3.0.6",
-    "husky": "^1.3.1",
-    "jest": "^27.3.1",
-    "jest-teamcity-reporter": "^0.9.0",
-    "lint-staged": "^13.0.3",
-    "mini-css-extract-plugin": "^2.5.2",
-    "mockdate": "^3.0.5",
-    "msw": "^0.48.2",
-    "msw-storybook-addon": "^1.6.3",
-    "node-polyfill-webpack-plugin": "^1.1.4",
-    "nodemon": "^2.0.15",
-    "prettier": "^2.5.1",
-    "raw-loader": "^4.0.2",
-    "react-test-renderer": "^17.0.2",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.4.4",
-    "utility-types": "^3.10.0",
-    "uuid": "^8.3.2",
-    "webpack": "^5.64.0",
-    "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.4.0",
-    "webpack-merge": "^5.8.0",
-    "webpack-node-externals": "^3.0.0",
-    "whatwg-fetch": "^2.0.4"
+    "cypress-plugin-stripe-elements": "1.0.2",
+    "eslint": "8.16.0",
+    "eslint-plugin-jest": "27.0.4",
+    "eslint-plugin-react": "7.30.0",
+    "fetch-mock": "9.11.0",
+    "git-revision-webpack-plugin": "3.0.6",
+    "husky": "1.3.1",
+    "jest": "27.4.5",
+    "jest-teamcity-reporter": "0.9.0",
+    "lint-staged": "13.0.3",
+    "mini-css-extract-plugin": "2.5.2",
+    "mockdate": "3.0.5",
+    "msw": "0.48.2",
+    "msw-storybook-addon": "1.6.3",
+    "node-polyfill-webpack-plugin": "1.1.4",
+    "nodemon": "2.0.15",
+    "prettier": "2.5.1",
+    "raw-loader": "4.0.2",
+    "react-test-renderer": "18.2.0",
+    "ts-jest": "27.1.2",
+    "typescript": "4.4.4",
+    "utility-types": "3.10.0",
+    "uuid": "8.3.2",
+    "webpack": "5.76.0",
+    "webpack-bundle-analyzer": "4.5.0",
+    "webpack-cli": "4.9.1",
+    "webpack-dev-server": "4.4.0",
+    "webpack-merge": "5.8.0",
+    "webpack-node-externals": "3.0.0",
+    "whatwg-fetch": "2.0.4"
   },
   "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@guardian/ab-core": "^2.0.0",
-    "@guardian/ab-react": "^2.0.1",
-    "@guardian/commercial-core": "^5.4.5",
-    "@guardian/consent-management-platform": "^12.0.0",
-    "@guardian/libs": "^14.0.0",
-    "@guardian/source-foundations": "^5.0.0",
-    "@guardian/source-react-components": "^5.0.0",
-    "@sentry/browser": "^5.22.3",
-    "@sentry/node": "^5.22.3",
-    "@stripe/react-stripe-js": "^1.1.2",
-    "@stripe/stripe-js": "^1.7.0",
+    "@emotion/react": "11.9.0",
+    "@guardian/ab-core": "2.0.0",
+    "@guardian/ab-react": "2.0.1",
+    "@guardian/consent-management-platform": "13.0.0",
+    "@guardian/libs": "14.0.0",
+    "@guardian/source-foundations": "10.0.1",
+    "@guardian/source-react-components": "13.0.0",
+    "@guardian/source-react-components-development-kitchen": "11.0.0",
+    "@sentry/browser": "5.22.3",
+    "@sentry/node": "5.22.3",
+    "@stripe/react-stripe-js": "1.16.5",
+    "@stripe/stripe-js": "1.52.1",
     "aws-sdk": "2.814.0",
-    "aws4": "^1.10.1",
-    "base-64": "^0.1.0",
-    "color": "^3.0.0",
-    "cookie-parser": "^1.4.4",
-    "csurf": "^1.10.0",
-    "date-fns": "^2.16.1",
-    "express": "^4.16.3",
-    "formik": "^2.2.9",
-    "helmet": "^3.23.3",
-    "js-cookie": "^2.2.1",
-    "lodash": "^4.17.21",
-    "node-fetch": "^2.6.7",
-    "number-to-words": "^1.2.4",
-    "ophan-tracker-js": "^1.3.16",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router": "^6.2.2",
-    "react-router-dom": "^6.2.2",
-    "regenerator-runtime": "^0.13.9",
-    "source-map-support": "^0.5.19",
-    "stylis-pxtorem": "^0.0.2",
-    "ua-parser-js": "^1.0.34",
-    "url-parse": "^1.5.9",
-    "winston": "^3.2.1",
-    "yup": "^0.28.2"
+    "aws4": "1.10.1",
+    "base-64": "0.1.0",
+    "color": "3.1.0",
+    "cookie-parser": "1.4.4",
+    "csurf": "1.10.0",
+    "date-fns": "2.16.1",
+    "express": "4.17.3",
+    "formik": "2.2.9",
+    "helmet": "3.23.3",
+    "js-cookie": "2.2.1",
+    "lodash": "4.17.21",
+    "node-fetch": "2.6.7",
+    "number-to-words": "1.2.4",
+    "ophan-tracker-js": "1.3.16",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router": "6.2.2",
+    "react-router-dom": "6.2.2",
+    "regenerator-runtime": "0.13.9",
+    "source-map-support": "0.5.21",
+    "stylis-pxtorem": "0.0.2",
+    "tslib": "2.5.0",
+    "url-parse": "1.5.9",
+    "winston": "3.2.1",
+    "yup": "0.28.5"
   },
   "license": "UNLICENSED",
   "msw": {
