@@ -3,15 +3,33 @@ import fetchMock from 'fetch-mock';
 import { ReactRouterDecorator } from '../../../../.storybook/ReactRouterDecorator';
 import { featureSwitches } from '../../../../shared/featureSwitches';
 import {
+	cancelledContribution,
+	cancelledGuardianWeekly,
+} from '../../../fixtures/cancelledProductDetail';
+import {
+	CancelledInAppPurchase,
+	InAppPurchaseIos,
+	PuzzleAppPurchaseAndroid,
+	PuzzleAppPurchaseIos,
+} from '../../../fixtures/inAppPurchase';
+import {
+	contributionCancelled,
 	contributionPayPal,
 	digitalDD,
+	guardianWeeklyCancelled,
 	guardianWeeklyCard,
+	guardianWeeklyGiftPurchase,
+	guardianWeeklyGiftRecipient,
 	newspaperVoucherPaypal,
 	supporterPlus,
+	supporterPlusCancelled,
 	toMembersDataApiResponse,
 } from '../../../fixtures/productDetail';
+import { singleContributionsAPIResponse } from '../../../fixtures/singleContribution';
 import { user } from '../../../fixtures/user';
 import { AccountOverview } from './AccountOverview';
+
+featureSwitches['appSubscriptions'] = true;
 
 export default {
 	title: 'Pages/AccountOverview',
@@ -26,6 +44,12 @@ export const NoSubscription: ComponentStory<typeof AccountOverview> = () => {
 	fetchMock
 		.restore()
 		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: { subscriptions: [] },
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		})
 		.get('/api/me/mma', { body: toMembersDataApiResponse() })
 		.get('/idapi/user', { body: user });
 
@@ -36,6 +60,12 @@ export const WithSubscriptions: ComponentStory<typeof AccountOverview> = () => {
 	fetchMock
 		.restore()
 		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: { subscriptions: [] },
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		})
 		.get('/api/me/mma', {
 			body: toMembersDataApiResponse(
 				guardianWeeklyCard,
@@ -47,25 +77,28 @@ export const WithSubscriptions: ComponentStory<typeof AccountOverview> = () => {
 	return <AccountOverview />;
 };
 
-export const WithContributionNewLayout: ComponentStory<
+export const WithContributionAndSwitchPossible: ComponentStory<
 	typeof AccountOverview
 > = () => {
-	featureSwitches['accountOverviewNewLayout'] = true;
-
 	fetchMock
 		.restore()
 		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: { subscriptions: [] },
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		})
 		.get('/api/me/mma', {
-			body: [contributionPayPal],
+			body: toMembersDataApiResponse(contributionPayPal),
 		});
 
 	return <AccountOverview />;
 };
 
-export const WithContributionNewLayoutPaymentFailure: ComponentStory<
+export const WithContributionInPaymentFailure: ComponentStory<
 	typeof AccountOverview
 > = () => {
-	featureSwitches['accountOverviewNewLayout'] = true;
 	const contributionPaymentFailure = {
 		...contributionPayPal,
 		alertText: 'Your payment has failed.',
@@ -74,6 +107,12 @@ export const WithContributionNewLayoutPaymentFailure: ComponentStory<
 	fetchMock
 		.restore()
 		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: { subscriptions: [] },
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		})
 		.get('/api/me/mma', {
 			body: toMembersDataApiResponse(
 				contributionPaymentFailure,
@@ -84,17 +123,113 @@ export const WithContributionNewLayoutPaymentFailure: ComponentStory<
 	return <AccountOverview />;
 };
 
-export const WithContributionNewLayoutDigisubAndContribution: ComponentStory<
+export const WithContributionAndSwitchNotPossible: ComponentStory<
 	typeof AccountOverview
 > = () => {
-	featureSwitches['accountOverviewNewLayout'] = true;
-
 	fetchMock
 		.restore()
 		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: { subscriptions: [] },
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		})
 		.get('/api/me/mma', {
-			body: [contributionPayPal, digitalDD],
+			body: toMembersDataApiResponse(contributionPayPal, digitalDD),
 		});
+
+	return <AccountOverview />;
+};
+
+export const WithCancelledSubscriptions: ComponentStory<
+	typeof AccountOverview
+> = () => {
+	fetchMock
+		.restore()
+		.get('/api/me/mma', {
+			body: toMembersDataApiResponse(
+				contributionCancelled,
+				guardianWeeklyCancelled,
+				supporterPlusCancelled,
+			),
+		})
+		.get('/api/cancelled/', {
+			body: [cancelledContribution, cancelledGuardianWeekly],
+		})
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: { subscriptions: [] },
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		});
+
+	return <AccountOverview />;
+};
+
+export const WithGiftSubscriptions: ComponentStory<
+	typeof AccountOverview
+> = () => {
+	fetchMock
+		.restore()
+		.get('/api/me/mma', {
+			body: toMembersDataApiResponse(
+				guardianWeeklyGiftRecipient,
+				guardianWeeklyGiftPurchase,
+			),
+		})
+		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: { subscriptions: [] },
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		});
+
+	return <AccountOverview />;
+};
+
+export const WithAppSubscriptions: ComponentStory<
+	typeof AccountOverview
+> = () => {
+	fetchMock
+		.restore()
+		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: {
+				subscriptions: [
+					CancelledInAppPurchase,
+					InAppPurchaseIos,
+					PuzzleAppPurchaseAndroid,
+					PuzzleAppPurchaseIos,
+				],
+			},
+		})
+		.get('/api/me/one-off-contributions', {
+			body: [],
+		})
+		.get('/api/me/mma', { body: toMembersDataApiResponse() })
+		.get('/idapi/user', { body: user });
+
+	return <AccountOverview />;
+};
+
+export const WithSingleContribution: ComponentStory<
+	typeof AccountOverview
+> = () => {
+	fetchMock
+		.restore()
+		.get('/api/cancelled/', { body: [] })
+		.get('/mpapi/user/mobile-subscriptions', {
+			body: {
+				subscriptions: [],
+			},
+		})
+		.get('/api/me/one-off-contributions', {
+			body: singleContributionsAPIResponse,
+		})
+		.get('/api/me/mma', { body: toMembersDataApiResponse() })
+		.get('/idapi/user', { body: user });
 
 	return <AccountOverview />;
 };
