@@ -8,11 +8,22 @@ import {
 	textSans,
 	until,
 } from '@guardian/source-foundations';
+import { capitalize } from 'lodash';
 import { Fragment } from 'react';
 import { parseDate } from '../../../../shared/dates';
+import { featureSwitches } from '../../../../shared/featureSwitches';
+import type {
+	AppSubscription,
+	MPAPIResponse,
+} from '../../../../shared/mpapiResponse';
+import {
+	AppStore,
+	determineAppStore,
+	isPuzzle,
+	isValidAppSubscription,
+} from '../../../../shared/mpapiResponse';
 import type {
 	InvoiceDataApiItem,
-	MembersDataApiItem,
 	MembersDataApiResponse,
 	PaidSubscriptionPlan,
 	ProductDetail,
@@ -21,7 +32,6 @@ import {
 	getMainPlan,
 	isGift,
 	isProduct,
-	mdapiResponseReader,
 	sortByJoinDate,
 } from '../../../../shared/productResponse';
 import type { GroupedProductTypeKeys } from '../../../../shared/productTypes';
@@ -31,7 +41,7 @@ import {
 	LoadingState,
 	useAsyncLoader,
 } from '../../../utilities/hooks/useAsyncLoader';
-import { allProductsDetailFetcher } from '../../../utilities/productUtils';
+import { allRecurringProductsDetailFetcher } from '../../../utilities/productUtils';
 import { GenericErrorScreen } from '../../shared/GenericErrorScreen';
 import { NAV_LINKS } from '../../shared/nav/NavConfig';
 import { EmptyAccountOverview } from '../accountoverview/EmptyAccountOverview';
@@ -57,8 +67,9 @@ type MMACategoryToProductDetails = {
 };
 
 type BillingResponse = [
-	MembersDataApiResponse | MembersDataApiItem[],
+	MembersDataApiResponse,
 	{ invoices: InvoiceDataApiItem[] },
+	MPAPIResponse,
 ];
 
 const subHeadingTitleCss = `
@@ -69,9 +80,9 @@ ${until.tablet} {
 };
 `;
 
-const subHeadingBorderTopCss = `
-border-top: 1px solid ${neutral['86']};
-margin: 50px 0 ${space[5]}px;
+const subHeadingBorderTopCss = css`
+	border-top: 1px solid ${neutral['86']};
+	margin: ${space[12]}px 0 ${space[5]}px;
 `;
 
 function decorateProductDetailWithInvoices(
@@ -79,11 +90,12 @@ function decorateProductDetailWithInvoices(
 ): ProductDetailWithInvoice {
 	return { ...productDetail, invoices: [] };
 }
+
 function joinInvoicesWithProductsInCategories(
-	mdapiObject: MembersDataApiResponse,
+	mdapiResponse: MembersDataApiResponse,
 	invoicesResponse: { invoices: InvoiceDataApiItem[] },
 ) {
-	const allProductDetails = mdapiObject.products
+	const allProductDetails = mdapiResponse.products
 		.filter(isProduct)
 		.sort(sortByJoinDate)
 		.map(decorateProductDetailWithInvoices);
@@ -296,6 +308,70 @@ function renderProductBillingInfo([mmaCategory, productDetails]: [
 	);
 }
 
+function getAppStoreMessage(subscription: AppSubscription) {
+	const iosMessage = 'Apple (for iOS)';
+	const androidMessage = 'Google (for Android)';
+
+	const appStore = determineAppStore(subscription);
+
+	switch (appStore) {
+		case AppStore.IOS:
+			return iosMessage;
+		case AppStore.ANDROID:
+			return androidMessage;
+		default:
+			return `${iosMessage}, or ${androidMessage}`;
+	}
+}
+
+function renderInAppPurchase(subscription: AppSubscription) {
+	const tableHeadingCss = css`
+		width: 100%;
+		${headline.xxsmall({ fontWeight: 'bold' })};
+		margin: 0;
+		padding: ${space[3]}px ${space[5]}px;
+		background-color: ${neutral[97]};
+		${until.tablet} {
+			font-size: 1.0625rem;
+			line-height: 1.6;
+			padding: ${space[3]}px;
+		}
+	`;
+	const puzzleOrNews = isPuzzle(subscription) ? 'puzzle' : 'news';
+
+	return (
+		<div css={subHeadingBorderTopCss} key={subscription.subscriptionId}>
+			<h2
+				css={css`
+					${subHeadingTitleCss}
+					margin: 0;
+				`}
+			>
+				{capitalize(puzzleOrNews)} app
+			</h2>
+			<div
+				css={css`
+					${textSans.medium()};
+					border: 1px solid ${neutral[86]};
+					display: flex;
+					flex-wrap: wrap;
+					margin: ${space[5]}px 0;
+				`}
+			>
+				<h2 css={tableHeadingCss}>Payment</h2>
+				<div
+					css={css`
+						padding: ${space[3]}px;
+					`}
+				>
+					To change your payment setup, please contact{' '}
+					{getAppStoreMessage(subscription)}.
+				</div>
+			</div>
+		</div>
+	);
+}
+
 function BillingDetailsComponent(props: {
 	mmaCategoryToProductDetails: MMACategoryToProductDetails;
 }) {
@@ -329,13 +405,18 @@ const BillingPage = () => {
 		return <GenericErrorScreen />;
 	}
 
-	const [mdapiResponse, invoicesResponse] = billingResponse;
-	const mdapiObject = mdapiResponseReader(mdapiResponse);
+	const [mdapiResponse, invoicesResponse, mpapiResponse] = billingResponse;
+	const appSubscriptions = mpapiResponse.subscriptions.filter(
+		isValidAppSubscription,
+	);
 
 	const { allProductDetails, mmaCategoryToProductDetails } =
-		joinInvoicesWithProductsInCategories(mdapiObject, invoicesResponse);
+		joinInvoicesWithProductsInCategories(mdapiResponse, invoicesResponse);
 
-	if (allProductDetails.length === 0) {
+	if (
+		(allProductDetails.length === 0 && appSubscriptions.length === 0) ||
+		(allProductDetails.length === 0 && !featureSwitches.appSubscriptions)
+	) {
 		return <EmptyAccountOverview />;
 	}
 
@@ -344,6 +425,9 @@ const BillingPage = () => {
 			<PaymentFailureAlertIfApplicable
 				productDetails={allProductDetails}
 			/>
+			{featureSwitches.appSubscriptions &&
+				appSubscriptions.map(renderInAppPurchase)}
+
 			<BillingDetailsComponent
 				mmaCategoryToProductDetails={mmaCategoryToProductDetails}
 			/>
@@ -353,8 +437,9 @@ const BillingPage = () => {
 
 const billingFetcher = () =>
 	Promise.all([
-		allProductsDetailFetcher(),
+		allRecurringProductsDetailFetcher(),
 		fetchWithDefaultParameters('/api/invoices'),
+		fetchWithDefaultParameters('/mpapi/user/mobile-subscriptions'),
 	]);
 
 export const Billing = () => {
