@@ -119,13 +119,8 @@ const ReasonPicker = ({
 							padding: ${space[5]}px;
 						`}
 					>
-						{productType.cancellation.reasons
-							.filter((reason: CancellationReason) =>
-								reason.shouldShow
-									? reason.shouldShow(productDetail)
-									: true,
-							)
-							.map((reason: CancellationReason) => (
+						{productType.cancellation.reasons.map(
+							(reason: CancellationReason) => (
 								<Radio
 									key={reason.reasonId}
 									name="cancellation-reason"
@@ -139,7 +134,8 @@ const ReasonPicker = ({
 										}
 									`}
 								/>
-							))}
+							),
+						)}
 					</RadioGroup>
 				</fieldset>
 				{inValidationErrorState && !selectedReasonId.length && (
@@ -296,6 +292,22 @@ interface ReasonPickerWithCancellationDateProps {
 	productDetail: ProductDetail;
 }
 
+function getChargedThroughDateStr(
+	cancellationDateResponse: CancellationDateResponse,
+) {
+	if (
+		cancellationDateResponse.cancellationEffectiveDate === 'now' ||
+		cancellationDateResponse.cancellationEffectiveDate === undefined ||
+		cancellationDateResponse.cancellationEffectiveDate === null
+	) {
+		return undefined;
+	}
+
+	return parseDate(
+		cancellationDateResponse.cancellationEffectiveDate,
+	).dateStr(DATE_FNS_LONG_OUTPUT_FORMAT);
+}
+
 const ReasonPickerWithCancellationDate = ({
 	productType,
 	productDetail,
@@ -329,9 +341,9 @@ const ReasonPickerWithCancellationDate = ({
 		return <GenericErrorScreen />;
 	}
 
-	const chargedThroughDateStr = parseDate(
-		cancellationDateResponse.cancellationEffectiveDate,
-	).dateStr(DATE_FNS_LONG_OUTPUT_FORMAT);
+	const chargedThroughDateStr = getChargedThroughDateStr(
+		cancellationDateResponse,
+	);
 
 	return (
 		<ReasonPicker
