@@ -1,13 +1,9 @@
 import { css } from '@emotion/react';
 import { space } from '@guardian/source-foundations';
-import {
-	Button,
-	Radio,
-	RadioGroup,
-	SvgArrowRightStraight,
-} from '@guardian/source-react-components';
+import { Button, Radio, RadioGroup } from '@guardian/source-react-components';
 import { useEffect, useState } from 'react';
 import type * as React from 'react';
+import { useNavigate } from 'react-router';
 import { trackEventInOphanOnly } from '../../../utilities/analytics';
 import { getGeoLocation } from '../../../utilities/geolocation';
 
@@ -27,6 +23,12 @@ const formContainerStyles = css`
 	}
 `;
 
+const buttonLayoutCss = css`
+	> * + * {
+		margin-left: ${space[3]}px;
+	}
+`;
+
 interface ReminderSignup {
 	reminderPeriod: string;
 	reminderOption: string;
@@ -38,7 +40,7 @@ interface ReminderChoice {
 	thankYouMessage: string;
 }
 
-const REMINDER_ENDPOINT = '/api/reminders';
+const REMINDER_ENDPOINT = '/api/reminders/create';
 const REMINDER_PLATFORM = 'MMA';
 const REMINDER_STAGE = 'WINBACK';
 const REMINDER_COMPONENT = 'CANCELLATION';
@@ -92,6 +94,8 @@ const getDefaultReminderChoices = (): ReminderChoice[] => [
 export const CancellationContributionReminder: React.FC = () => {
 	const [selectedChoiceIndex, setSelectedChoiceIndex] = useState(0);
 	const [hasSetReminder, setHasSetReminder] = useState(false);
+
+	const navigate = useNavigate();
 
 	const email = window.guardian.identityDetails.email;
 	const reminderChoices = getDefaultReminderChoices();
@@ -163,14 +167,15 @@ export const CancellationContributionReminder: React.FC = () => {
 								/>
 							))}
 						</RadioGroup>
-
-						<Button
-							onClick={onSubmit}
-							icon={<SvgArrowRightStraight />}
-							iconSide="right"
-						>
-							Set my reminder
-						</Button>
+						<div css={buttonLayoutCss}>
+							<Button onClick={onSubmit}>Set my reminder</Button>
+							<Button
+								priority="tertiary"
+								onClick={() => navigate('/')}
+							>
+								Back to your account
+							</Button>
+						</div>
 					</div>
 				</div>
 			)}
