@@ -3,13 +3,11 @@ import { createContext, useState } from 'react';
 import { Navigate, Outlet, useLocation } from 'react-router-dom';
 import type { DateRange } from '../../../../shared/dates';
 import type {
-	MembersDataApiItem,
 	MembersDataApiResponse,
 	ProductDetail,
 } from '../../../../shared/productResponse';
 import {
 	isProduct,
-	mdapiResponseReader,
 	MembersDataApiAsyncLoader,
 } from '../../../../shared/productResponse';
 import type {
@@ -99,10 +97,8 @@ const handleMembersDataResponse =
 		publicationsImpacted: HolidayStopDetail[],
 		setPublicationsImpacted: Dispatch<SetStateAction<HolidayStopDetail[]>>,
 	) =>
-	(data: [MembersDataApiResponse | MembersDataApiItem[]]) => {
-		const mdaResponse = mdapiResponseReader(data);
-
-		const filteredProductDetails = mdaResponse.products.filter(isProduct);
+	(mdapiResponse: MembersDataApiResponse) => {
+		const filteredProductDetails = mdapiResponse.products.filter(isProduct);
 
 		if (filteredProductDetails.length === 1) {
 			const productDetail = filteredProductDetails[0];
