@@ -1,13 +1,12 @@
-import { css } from '@emotion/react';
+import type { SerializedStyles } from '@emotion/react';
 import type { FC } from 'react';
 import { WithStandardTopMargin } from '../../../shared/WithStandardTopMargin';
 import { Lines } from '../Lines';
 import { MarketingToggle } from '../MarketingToggle';
 import type { ConsentOption } from '../models';
 import { PageSection } from '../PageSection';
-import { standardSansText, toggleDescriptionPadding } from '../sharedStyles';
 
-type ClickHandler = (id: string) => {};
+export type ClickHandler = (id: string) => {};
 
 interface OptOutSectionProps {
 	consents: ConsentOption[];
@@ -29,10 +28,11 @@ const consentSubscribedValueInverter = (
 	};
 };
 
-const optOutFinder =
+export const optOutFinder =
 	(
 		consents: ConsentOption[],
 		clickHandler: ClickHandler,
+		customCSS?: SerializedStyles,
 		invertSubscribedValue?: (c: ConsentOption) => ConsentOption,
 	) =>
 	(id: string) => {
@@ -49,61 +49,11 @@ const optOutFinder =
 					description={consent.description} // Not all consents from IDAPI have a description
 					selected={consent.subscribed}
 					onClick={clickHandler}
+					divCss={customCSS}
 				/>
 			)
 		);
 	};
-
-const yourDataStyles = css`
-	ul {
-		padding-inline-start: 20px;
-		margin-block-start: 0;
-		line-height: 1.5;
-	}
-
-	p {
-		margin-bottom: 6px;
-	}
-`;
-
-// NOTE: The Your Data description is hard coded here and in
-// Identity Gateway. They should be the same unless running A/B tests.
-const YourDataDescription: FC = () => (
-	<div css={yourDataStyles}>
-		<p>What we mean by your data:</p>
-		<ul>
-			<li>Information you provide e.g. email address</li>
-			<li>Products or services you buy from us</li>
-			<li>
-				Pages you view on theguardian.com or other Guardian websites
-				when signed in and where you are reading from
-			</li>
-		</ul>
-	</div>
-);
-
-// NOTE: The personalised advertising consent description is hard coded here and in
-// Identity Gateway. They should be the same unless running A/B tests.
-const PersonalisedAdvertisingConsentDescription: FC = () => (
-	<div css={[standardSansText, toggleDescriptionPadding, yourDataStyles]}>
-		<p>
-			Advertising is a crucial source of our funding. You won't see more
-			ads, and your data won't be shared with third parties to use for
-			their own advertising.
-		</p>
-		<p>We do this by:</p>
-		<ul>
-			<li>
-				Analysing your information to predict what you might be
-				interested in.
-			</li>
-			<li>
-				Checking if you are already a customer of other trusted
-				partners.
-			</li>
-		</ul>
-	</div>
-);
 
 export const OptOutSection: FC<OptOutSectionProps> = (props) => {
 	const { consents, clickHandler } = props;
@@ -111,9 +61,9 @@ export const OptOutSection: FC<OptOutSectionProps> = (props) => {
 	const addInvertedMarketingToggle = optOutFinder(
 		consents,
 		clickHandler,
+		undefined,
 		consentSubscribedValueInverter,
 	);
-	const addMarketingToggle = optOutFinder(consents, clickHandler);
 
 	return (
 		<>
@@ -130,16 +80,6 @@ export const OptOutSection: FC<OptOutSectionProps> = (props) => {
 			</PageSection>
 			<WithStandardTopMargin>
 				<Lines n={1} />
-			</WithStandardTopMargin>
-			<WithStandardTopMargin>
-				<PageSection
-					title="Your data"
-					description={<YourDataDescription />}
-				>
-					{addInvertedMarketingToggle('profiling_optout')}
-					{addMarketingToggle('personalised_advertising')}
-					<PersonalisedAdvertisingConsentDescription />
-				</PageSection>
 			</WithStandardTopMargin>
 		</>
 	);
