@@ -1,46 +1,11 @@
 import { css } from '@emotion/react';
-import { from, palette, space, textSans } from '@guardian/source-foundations';
+import { space } from '@guardian/source-foundations';
 import { SvgTickRound } from '@guardian/source-react-components';
 import { useState } from 'react';
 import { expanderButtonCss } from '../../shared/ExpanderButton';
-
-const benefitsButtonCss = css`
-	${textSans.small()}
-	margin-top: ${space[1]}px;
-	padding: 0;
-	color: ${palette.brand[500]};
-	border-bottom: 1px solid ${palette.brand[500]};
-`;
+import { benefitsButtonCss, benefitsCss, lineBreakCss } from './BenefitsStyles';
 
 export const SupporterPlusBenefitsSection = () => {
-	const benefitsCss = css`
-		${textSans.medium()};
-		list-style: none;
-		margin: 0 0 0 -4px;
-		padding: 0;
-
-		li + li {
-			margin-top: ${space[2]}px;
-		}
-
-		li {
-			display: flex;
-			align-items: flex-start;
-		}
-
-		svg {
-			flex-shrink: 0;
-			margin-right: ${space[2]}px;
-			fill: ${palette.brand[500]};
-		}
-	`;
-
-	const lineBreakCss = css`
-		${from.tablet} {
-			display: none;
-		}
-	`;
-
 	return (
 		<ul id="benefits" css={benefitsCss}>
 			<li>
