@@ -1,3 +1,5 @@
+import { loadScript } from '@guardian/libs';
+import * as Sentry from '@sentry/browser';
 import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import parse from 'url-parse';
@@ -15,12 +17,14 @@ declare global {
 	}
 }
 
+const GA_AVAILABLE = false;
 const GA_UA = 'UA-51507017-5';
 
 export const useAnalytics = () => {
 	const location = useLocation();
 	const [cmpIsInitialised, setCmpIsInitialised] = useState<boolean>(false);
 	const [gaIsInitialised, setGaIsInitialised] = useState<boolean>(false);
+	const [qmIsInitialised, setQmIsInitialised] = useState<boolean>(false);
 
 	const initialiseGa = () => {
 		// Run self evoking GA script
@@ -76,6 +80,24 @@ export const useAnalytics = () => {
 		setGaIsInitialised(true);
 	};
 
+	const initialiseQm = () => {
+		loadScript(
+			'https://cdn.quantummetric.com/instrumentation/1.32.18/quantum-gnm.js',
+			{
+				async: true,
+				integrity:
+					'sha384-JB9+fEddky647lwDnP+fbHslLQq+tyL6ozahvpgl4fxgCWjU5e6qxU8FIcCYiosU',
+				crossOrigin: 'anonymous',
+			},
+		)
+			.then(() => {
+				setQmIsInitialised(true);
+			})
+			.catch(() => {
+				Sentry.captureException('Failed to load Quantum Metric');
+			});
+	};
+
 	useEffect(() => {
 		import('@guardian/consent-management-platform').then(
 			({ onConsentChange, getConsentFor }) => {
@@ -85,11 +107,17 @@ export const useAnalytics = () => {
 						consentState,
 					);
 
+					const qmConsentState = getConsentFor('qm', consentState);
+
 					// @ts-expect-error: Suppressing "element implicitly has an 'any' type because index expression is not of type 'number'."
 					window[`ga-disable-${GA_UA}`] = !gaConsentState;
 
-					if (gaConsentState && !gaIsInitialised) {
+					if (GA_AVAILABLE && gaConsentState && !gaIsInitialised) {
 						initialiseGa();
+					}
+
+					if (qmConsentState && !qmIsInitialised) {
+						initialiseQm();
 					}
 
 					setCmpIsInitialised(true);
@@ -123,7 +151,6 @@ export const useAnalytics = () => {
 				dimension12: window.guardian.INTCMP,
 				dimension29: MMA_AB_TEST_DIMENSION_VALUE,
 			});
-			// TODO add ophan pageViewId as a GA dimension
 			applyAnyOptimiseExperiments();
 		}
 	}, [location, cmpIsInitialised, gaIsInitialised]);
