@@ -1,13 +1,15 @@
-import { contributionCard } from '../../../client/fixtures/productDetail';
+import {
+	contributionCard,
+	toMembersDataApiResponse,
+} from '../../../client/fixtures/productDetail';
 import { signInAndAcceptCookies } from '../../lib/signInAndAcceptCookies';
 
 describe('Update contribution amount', () => {
 	const setSignInStatus = () => {
 		cy.window().then((window) => {
-			// @ts-ignore
 			window.guardian.identityDetails = {
 				signInStatus: 'signedInRecently',
-				userId: '200006712',
+				userId: '1',
 				displayName: 'user',
 				email: 'example@example.com',
 			};
@@ -21,13 +23,23 @@ describe('Update contribution amount', () => {
 
 		cy.intercept('GET', '/api/me/mma?productType=Contribution', {
 			statusCode: 200,
-			body: [contributionCard],
+			body: toMembersDataApiResponse(contributionCard),
 		});
 
 		cy.intercept('GET', '/api/me/mma', {
 			statusCode: 200,
-			body: [contributionCard],
+			body: toMembersDataApiResponse(contributionCard),
 		});
+
+		cy.intercept('GET', '/mpapi/user/mobile-subscriptions', {
+			statusCode: 200,
+			body: { subscriptions: [] },
+		});
+
+		cy.intercept('GET', '/api/me/one-off-contributions', {
+			statusCode: 200,
+			body: [],
+		}).as('single_contributions');
 
 		cy.intercept('GET', '/api/cancelled/', {
 			statusCode: 200,
