@@ -24,7 +24,7 @@ export const initFeatureSwitchUrlParamOverride = () => {
  */
 export const featureSwitches: Record<string, boolean> = {
 	exampleFeature: false,
-	cancellationProductSwitch: false,
-	accountOverviewNewLayout: false,
-	productSwitching: true,
+	appSubscriptions: true,
+	singleContributions: false,
+	membershipSave: false,
 };
